include: add _NL_NO_WARN_DEPRECATED_HEADER for suppressing warning about deprecated headers
Header files are part of a stable API. Warning about using deprecated
API is cumbersome, albeit often useful to force the user to migrate
away from the API.

But it also requires that the unit test disables "-Wcpp". That is
problematic on its own, because we don't want to disable any warnings.

As we only have 4 such header files, instead guard the warnings with
a

  #ifndef _NL_NO_WARN_DEPRECATED_HEADER