github,clang-format: update fedora version for clang-format
We use clang-format, but the result depends on the version of the tool.
The version to be used, is the one from the github action. You can use
"tools/clang-format-container.sh" to run that version in a container.

Now that Fedora 40 is released, bump the Fedora version (and the used
clang-format version) from Fedora 39 (17.0.6-2.fc39) to Fedora 40
(18.1.1-1.fc40). Luckily, our source code formats the same for both.