route: use the new helper function for comparing nexthops
When a route is created while the interface has no link, we get a
notification with the route and the nexthop having the flag LINKDOWN.

If the interface later gets a link, we do not get a route notification
about it, so the route and nexthop stay at LINKDOWN in the libnl cache.

If the route then gets removed again, the to be removed route will not
have the LINKDOWN flag anymore, which then can break comparison of the
nexthop(s).

So use the new nexthop identical helper to avoid this scenario.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>