build: move "-DPGKLIBDIR" and rename
These defines should be defined at one place ("defines_cppflags")
and they should be available the same throughout the source code.

Even if PKGLIBDIR is only used by "src/lib/utils.c", there is no harm
in defining it for all compilation units. The point is that it is only
defined at one place and the same for all compilation units.