core: support receiving multipart objects
In order to support an upcoming patch implementing support for
requesting bridge vlan information for a single interface, a
modification to the pickup code is needed to allow handling of multipart
messages.

Based on the implementation of the bridge vlan dumping code in the
kernel, information relating to a single interface may not all be
contained within a single message. If the maximum message size is
exceeded before all entries can be filled in, a new message will be
started with the same ifindex but containing the remaining entries.

The code is currently only designed for handling netlink responses where
the provided parser function will be called a maximum of one time. This
is based on the assumption that all data relating to an object will be
contained within a single response message. If multiple messages are
received and the parser function is invoked multiple times, previously
stored objects will be leaked and the resulting object will only contain
information from the last parsed message.

In order to handle the case where multiple messages need to be parsed to
construct a single object, use the already existing oo_update function
that may be defined for an object. The function will be called to merge
partial objects (that are the result of each invocation of the parser
function) together into a single object that is returned at the end.