build: use "-std=gnu11" as compiler flags and depend on gcc's C11
So far, we didn't specify "-std" option, and it was ill defined which C
behavior we require.

Note that we already use __attribute__((__cleanup__(fcn))), so
effectively we require either gcc or clang. A standard C compiler won't
suffice. That is for good reasons, because features like the cleanup
attribute are IMO invaluable and we should not write "C" without it.

Also, it's 2024. It's basically impossilbe that you use gcc/clang with a
version that doesn't yet support (gcc's) C11 variant. So explicitly
require that via the "-std=gnu11" option. I don't think we currently
require newer C features, so this is sufficient for us and for now.