clang-format: bump clang-format version to latest Fedora 41
We use clang-format for formatting the code. The actual formatting
depends on the version of clang-format, as the tool improves, the
formatting slightly changes. We thus need to agree on which version to
use.

The used version is the one that gets check in the github test, and it's
the one from a recent Fedora release.

Now, with Fedora 41 out, bump the version from 18.1.8-1.fc40 to
19.1.0-1.fc41.