[cache] Adopt resizeable hashtable
First, get rid of co_hash_size. It is used nowhere and the adoption of
resizeable hashtables will make that argument obsolete.

Adoption of resizeable hashtables significantly improves the time it takes to
create & populate the cache. For example, for 500k ip nexthops, `nl-nh-list`
call goes from ~12s to ~3s.

Signed-off-by: Christoph Paasch <cpaasch@openai.com>