[ip6gre/ip6vti]: Fix printing of the link-name
When _ATTR_LINK is set, the link needs to be looked up before printing.
Similar as is done for ip6tnl.

So, we need to fix this in ip6gre and ip6vti.

TODO - it looks like we are leaking parent here.... I guess I need to
nl_auto free it... to be confirmed!!!

Signed-off-by: Christoph Paasch <cpaasch@openai.com>