[link] Avoid leaking rtnl_link object
When dumping and the *_ATTR_LINK flag is present, we do a call to
link_lookup() which ends up getting a references on the link object.

Thus, if we want to avoid leaking things, we need to actually
rtnl_link_put the object. Do this with _nl_auto_rtnl_link.

Signed-off-by: Christoph Paasch <cpaasch@openai.com>