nl-base-utils: declare static buffer size in _nl_inet_ntop()
For _nl_inet_ntop() we either need at least a buffer of size INET_ADDRSTRLEN or
INET6_ADDRSTRLEN, depending on the address family. We cannot express to
the compiler which one it should be, but it certainly needs to be at
least INET_ADDRSTRLEN.

While not perfect, I think it's better to tell the compiler we need at
least this many bytes. We might get a better compiler warning for bugs.

This is also why you usually should prefer _nl_inet_ntop4() or
_nl_inet_ntop6().