all: use _nl_inet_ntop() helpers
When we call inet_ntop() with a AF_INET/AF_NET6 address family,
we almost always will provide a suitably large buffer. In those
cases, the libc function has no reason to ever fail. Handling
errors there is not useful. Instead, use _nl_inet_ntop() which
asserts that it does not fail and does not require us to handle
errors.