error: revert mapping EHOSTUNREACH/ENETDOWN
Earlier commits

  - 45449e0bc693 ('Add EHOSTUNREACH to nl_syserr2nlerr')
  - fe74608d550e ('Add ENETDOWN to nl_syserr2nlerr')

introduced NLE_HOSTUNREACH and NLE_NETDOWN. So now failures with those
error codes would no longer return NLE_FAILURE. This is a desirable
change (in principle), but also an API break.

As it causes problems ([1]), gut the change for now. We keep the error
defines (and the assigned NLE error numbers), but we no longer map the
errno.h codes, which in practice reverts those two changes.

This is only an intermediate solution, to revert the break. We should
bring this change back, but somehow "better". It's not clear how exactly
we can improve the API without breaking it. For example, we could
identify a few old functions which must stick to the old behavior, while
new functions get the new error codes. Or maybe better, selected
functions should guarantee that when a function returns NLE_FAILURE that
errno is set to a value that is relevant to the underlying.  failure.

[1] https://github.com/thom311/libnl/pull/433#issuecomment-3625794675